organized files into proper folders
