+++ src/Components/OrgEvents/updateEvent.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 
 function UpdateEvent() {
   const { id } = useParams();
